requirements.txt
torch
torchvision
transformers
gradio==4.44.1
numpy
opencv-python-headless
Pillow